refactor: 컴포넌트명 video -> preview로 수정
